(refactor) Create VobPfxRenderer + (feat) create StaticCache for oCVob (Pfx+Decal+Mesh).
